requirements.txt
Django>=4.2
moviepy
python-docx
openai-whisper
torch
ffmpeg-python
gunicorn==21.2.0
